3.1.12 node->stacklist
